[#69] chore: svg move.
